Git Bash commit update

+++ LYR/LYRLog.sh
@@ -317,19 +317,30 @@ function StartLogFile { # (AFileName: str):
     LOG_OPT_2=${LOG_OPT:1:1}
     echo "LOG_OPT_1=$LOG_OPT_1"
     echo "LOG_OPT_2=$LOG_OPT_2"
-
+    if [[ "$LOG_OPT_1 -eq '1'" ]] ; then
+        LOG_FILE_ADD=1
+    else
+        LOG_FILE_ADD=0
+    fi
+    if [[ "$LOG_OPT_2 -eq '1'" ]] ; then
+        LOG_FILE_DT=1
+    else
+        LOG_FILE_DT=0
+    fi
     echo "LOG_FILE_ADD=$LOG_FILE_ADD"
     echo "LOG_FILE_DT=$LOG_FILE_DT"
     
     # 
     LOG_STR=""
     
-    LFileName="$LOG_FILE"
-    if [ -r "$LFileName" ] ; then
-        echo "$LFileName"
-        rm "$LFileName"
+    if [[ "$LOG_FILE_ADD -eq 1" ]] ; then
+        LFileName="$LOG_FILE"
+        if [ -r "$LFileName" ] ; then
+            echo "$LFileName"
+            rm "$LFileName"
+        fi
+        touch "$LFileName"
     fi
-    touch "$LFileName"
     exec 3>>"$LFileName"
     # -------------------------------------------------------------------
     AddLog $loAll $INFO "Старт: $(date +"$FORMAT")"
